requirements.txt
allure-python-commons
pytest
selenium
webdriver-manager
